move `get_symbol_list` to task_retrieval
it's used in several places
