+++ scripts/tasks_retrieval/Reactome_tasks_creation.py
@@ -3,6 +3,7 @@ import pandas as pd
 import requests
 
 from gene_benchmark.task_retrieval import (
+    get_symbol_list,
     list_form_to_onehot_form,
     verify_source_of_data,
 )
@@ -26,27 +27,6 @@ def get_token_link_for_symbols(symbols: list[str]) -> str:
     """
     token = get_token(symbols)
     return f"https://reactome.org/AnalysisService/download/{token}/pathways/TOTAL/result.csv"
-
-
-def get_symbol_list(
-    url: str = "https://g-a8b222.dd271.03c0.data.globus.org/pub/databases/genenames/hgnc/json/hgnc_complete_set.json",
-) -> list[str]:
-    """
-    Retrieves the symbol list from a HGNC json like file.
-
-    Args:
-    ----
-        url (str, optional): url for the json file download. Defaults to "https://g-a8b222.dd271.03c0.data.globus.org/pub/databases/genenames/hgnc/json/hgnc_complete_set.json".
-
-    Returns:
-    -------
-        list[str]: list of symbols
-
-    """
-    with requests.get(url) as response:
-        response.raise_for_status()
-        reactome_res = response.json()
-    return [v["symbol"] for v in reactome_res["response"]["docs"]]
 
 
 def get_token(
